requirements.txt
flask
line-bot-sdk
anthropic
gunicorn
pandas
apscheduler
pyngrok
requests
python-dotenv
